EXA-22486 | Restrict the deletion of claims if claim has payment and show correct warning message

Approved-by: Deryck Henson
Approved-by: Adam Gillfillan
Approved-by: Nirmilla Roy

+++ modules/ahs/index.js
@@ -199,16 +199,24 @@ const ahsmodule = {
 
         const claimDeleteAccess = await ahs.getPendingTransactionCount(args);
         const {
-            pending_transaction_count = null,
+            pending_transaction_count = 0,
             payment_entry_count = 0,
-            claim_balance_amount = null,
-            claim_total_amount = null
+            claim_balance_amount = 0,
+            claim_total_amount = 0,
+            claim_applied = 0,
+            claim_adjustment = 0
         } = claimDeleteAccess.rows && claimDeleteAccess.rows[0] || {};
 
-        let allowDelete = (payment_entry_count > 0 && (claim_balance_amount == claim_total_amount) && claimStatusCode === 'D')
+        let allowDelete = (payment_entry_count > 0 && (claim_balance_amount === claim_total_amount) && claimStatusCode === 'D')
                             || (payment_entry_count == 0 && ['R', 'BR'].indexOf(claimStatusCode) !== -1)
                             || (pending_transaction_count == 0 && payment_entry_count == 0);
 
+        if (!allowDelete && claim_applied !== 0 && claim_adjustment !== 0) {
+            return {
+                message: 'Claim has payment, Please unapply before delete'
+            };
+        }
+        
         if (allowDelete) {
             const claimDeletedResult = await ahs.purgeClaim(args);
             const {
@@ -216,7 +224,7 @@ const ahsmodule = {
             } = claimDeletedResult.rowCount && claimDeletedResult.rows[0] || {};
 
             return {
-                message: purge_claim_or_charge ? 'Claim deleted successfully' : 'Could not delete claim, AHS response pending',
+                message: 'Claim deleted successfully',
                 isClaimDeleted: purge_claim_or_charge
             };
